requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
args==0.1.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backrefs==5.8
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clint==0.5.1
colorama==0.4.6
contourpy==1.3.1
coverage==7.8.0
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
deptry==0.23.0
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.13
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.44
griffe==1.7.2
grpcio==1.71.0
huggingface-hub==0.30.2
hydra-core==1.3.2
identify==2.6.9
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.1.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.34.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
line-profiler==4.2.0
loguru==0.7.3
markdown==3.7
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.11
mkdocs-material-extensions==1.3.1
mkdocstrings==0.29.1
mkdocstrings-python==1.16.8
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
paginate==0.5.7
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peft==0.15.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==3.1.1
pre-commit==4.2.0
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.2
pydantic-core==2.33.1
pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.3
pyproject-api==1.9.0
pytest==7.2.0
pytest-cov==6.1.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2025.2
pyyaml==6.0.2
pyyaml-env-tag==0.1
regex==2024.11.6
requests==2.32.3
requirements-parser==0.11.0
ruff==0.11.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.4.1
torchmetrics==1.7.1
torchsde==0.2.6
torchvision==0.19.1
tox==4.25.0
tox-uv==1.25.0
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.51.1
triton==3.0.0
types-setuptools==78.1.0.20250329
typing-extensions==4.13.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uv==0.6.12
virtualenv==20.30.0
wandb==0.19.9
watchdog==6.0.0
wcwidth==0.2.13
webdataset==0.2.111
werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.19.0
zipp==3.21.0
